fixed dan's code (kinda)
